MCS-642 WIP of moving output handling outside of controller.  I think it works.  Tests definitely won't
